Adding basic todo detail view. Refactoring needed to correct names.
Basic todo detail view now supports showing the list of todo items, and showing if each item is checked or not. Does not yet update the todo item if the checked state is changed.

The count of todo items is correct, however it does not yet count the selected items.