Removing the todo item detail activity and showing all fragments from a single main activity.
This removes the unneeded second activity and makes the code much simpler for managing the todo items in memory.

Also this means that any changes made to a TodoItem or TodoGroup are immediately reflected in all the other views when they are re-drawn as they all deal with the same in-memory copy of the object. This is also much more memory efficient than serializing objects between views -- creating copies.